Rename Command to AbstractCommand
